# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  invoice_comments.CreateCommentDTO:
    properties:
      comment_text:
        maxLength: 2000
        minLength: 1
        type: string
      invoice_id:
        type: string
      parent_id:
        type: string
      token:
        type: string
    required:
    - comment_text
    - invoice_id
    type: object
  invoice_comments.ReactToCommentDTO:
    properties:
      email:
        type: string
      emoji:
        maxLength: 10
        minLength: 1
        type: string
      token:
        type: string
    required:
    - emoji
    type: object
  invoice_comments.UpdateCommentDTO:
    properties:
      comment_text:
        maxLength: 2000
        minLength: 1
        type: string
      email:
        type: string
      token:
        type: string
    required:
    - comment_text
    type: object
...
  utils.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:4000
info:
  contact:
    email: davidhero125@gmail.com
    name: API Support
  description: Invoice management and payment platform API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Stellance API
  version: "1.0"
paths:
...
  /comments:
    get:
      parameters:
      - description: Invoice ID
        in: query
        name: invoice_id
        required: true
        type: string
      - description: Parent comment ID (for replies)
        in: query
        name: parent_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: ASC or DESC
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: List comments for an invoice
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Comment data (include token for guest access)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/invoice_comments.CreateCommentDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Post a comment on an invoice
      tags:
      - comments
  /comments/{id}:
    get:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Get comment by ID
      tags:
      - comments
    patch:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated comment text
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/invoice_comments.UpdateCommentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Edit a comment
      tags:
      - comments
  /comments/{id}/react:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Emoji to remove
        in: query
        name: emoji
        required: true
        type: string
      - description: Guest email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Remove emoji reaction from a comment
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Emoji and optional guest token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/invoice_comments.ReactToCommentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Add emoji reaction to a comment
      tags:
      - comments
...
  /notifications:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: count
        type: integer
      - description: ASC or DESC
        in: query
        name: order_by
        type: string
      - description: Filter by viewed status
        in: query
        name: viewed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user notifications
      tags:
      - notifications
  /notifications/{id}:
    delete:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete a notification
      tags:
      - notifications
    get:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get notification by ID
      tags:
      - notifications
    patch:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      - description: Viewed status
        in: query
        name: viewed
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as viewed/unviewed
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"